refactor(xtask): split update_status.rs into per-subsystem modules (#4174)
Converts the 1383-LOC monolithic xtask/src/tasks/update_status.rs into a
directory module with six subsystem-owned files. Concurrent builders now edit
disjoint files instead of overlapping regions of one large file.

File structure after split:
- xtask/src/tasks/update_status/mod.rs     (dispatcher, shared helpers, enum)
- xtask/src/tasks/update_status/lsp.rs     (LSP coverage + compliance table)
- xtask/src/tasks/update_status/tests.rs   (test counts + missing-docs)
- xtask/src/tasks/update_status/parser.rs  (corpus tracking + sweep reports)
- xtask/src/tasks/update_status/quality.rs (mutation + crate table + UX receipt)
- xtask/src/tasks/update_status/dap.rs     (DAP test counts)
- xtask/src/tasks/update_status/workspace.rs (workspace fixtures + SLO)

Public API (StatusSubsystem, run()) is unchanged. All existing tests pass.
No file exceeds 400 LOC. Adds structural test verifying the split exists.
